refactor(tokenizer): move stemmer tests
